+++ Trader.cpp
@@ -80,8 +80,8 @@ void Trader::buyStock(Stock *stock, int amount){
     double accountBalance = this->getBalance();
     double purchasePrice = (stock->getPrice() * amount);
 
-    int actualPurchasePrice = (int)purchasePrice * 100;
-    int actualAccountBalance = (int)accountBalance * 100;
+    unsigned long actualPurchasePrice = (unsigned long)purchasePrice * 100;
+    unsigned long actualAccountBalance = (unsigned long)accountBalance * 100;
 
     if (actualAccountBalance - actualPurchasePrice == 0) {
         this->setBalance(0);
